requirements.txt
streamlit>=1.32.0
openai>=1.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.30